HMAC based One Time Password (HOTP)

An R6 class that implements the HMAC based One Time Password (HOTP) algorithm.

Initialization

HOTP$new(secret, digits = 6L, algorithm = "sha1")

Create an One Time Password object

secret a scalar character, the base32-based secret key

digits an integer, the number of digits of the password

algorithm the hash algorithm used, possible values are "sha1", "sha256" and "sha512".

Methods

HOTP$at(counter)

Generate an one time password at counter value

counter a non-negative integer

See also

Examples

p <- HOTP$new("")