Abstract R6 Class for One Time Password

Methods

Public methods

OTP$new()

OTP$provisioning_uri()

OTP$print()

OTP$clone()

Method new()

Create an One Time Password object

Usage

OTP$new(secret, digits = 6L, algorithm = "sha1")

Arguments

secret

a base32 based secret key

digits

an integer, the number of digits of the password

algorithm

the hash algorithm used, possible values are "sha1", "sha256" and "sha512".

Method provisioning_uri()

Internal use only

Usage

OTP$provisioning_uri(type, name, ...)

Arguments

type

internal use only

name

internal use only

...

internal use only

Method print()

print method

Usage

OTP$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

OTP$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.